Reogranize into a src dir, start moving things into logical places
