+++ src/can2040-xdrive-servo.c
@@ -9,6 +9,9 @@
 #include <string.h>
 
 #include "can2040/src/can2040.h"
+#include "pico-oled-driver/picoOled.h"
+#include "canbus.h"
+#include "xdrive_canbus.h"
 
 // Define the ADC channels for the joystick axes
 #define JOYSTICK_X_ADC 0
@@ -18,143 +21,6 @@
 // ADC0 is on GPIO26, ADC1 is on GPIO27
 #define JOYSTICK_X_PIN 26
 #define JOYSTICK_Y_PIN 27
-
-// ODrive CAN Commands
-#define ODR_CMD_HEARTBEAT           0x001
-#define ODR_CMD_ESTOP               0x002
-#define ODR_CMD_GET_MOTOR_ERROR     0x003
-#define ODR_CMD_GET_ENCODER_ERROR   0x004
-#define ODR_CMD_GET_SENSORLESS_ERROR 0x005
-#define ODR_CMD_SET_AXIS_NODE_ID    0x006
-#define ODR_CMD_SET_AXIS_STATE      0x007
-#define ODR_CMD_GET_ENCODER_ESTIMATES 0x009
-#define ODR_CMD_GET_ENCODER_COUNT   0x00A
-#define ODR_CMD_SET_CONTROLLER_MODES 0x00B
-#define ODR_CMD_SET_INPUT_POS       0x00C
-#define ODR_CMD_SET_INPUT_VEL       0x00D
-#define ODR_CMD_SET_INPUT_TORQUE    0x00E
-#define ODR_CMD_SET_VEL_LIMIT       0x00F
-#define ODR_CMD_START_ANTICOGGING   0x010
-#define ODR_CMD_SET_TRAJ_VEL_LIMIT  0x011
-#define ODR_CMD_SET_TRAJ_ACCEL_LIMITS 0x012
-#define ODR_CMD_SET_TRAJ_INERTIA    0x013
-#define ODR_CMD_GET_IQ              0x014
-#define ODR_CMD_GET_SENSORLESS_ESTIMATES 0x015
-#define ODR_CMD_REBOOT              0x016
-#define ODR_CMD_GET_VBUS_VOLTAGE    0x017
-#define ODR_CMD_CLEAR_ERRORS        0x018
-#define ODR_CMD_SET_LINEAR_COUNT    0x019
-#define ODR_CMD_SET_POS_GAIN        0x01A
-#define ODR_CMD_SET_VEL_GAINS       0x01B
-
-// ODrive Axis States
-#define AXIS_STATE_UNDEFINED                    0
-#define AXIS_STATE_IDLE                         1
-#define AXIS_STATE_STARTUP_SEQUENCE             2
-#define AXIS_STATE_FULL_CALIBRATION_SEQUENCE    3
-#define AXIS_STATE_MOTOR_CALIBRATION            4
-#define AXIS_STATE_SENSORLESS_CONTROL           5
-#define AXIS_STATE_ENCODER_INDEX_SEARCH         6
-#define AXIS_STATE_ENCODER_OFFSET_CALIBRATION   7
-#define AXIS_STATE_CLOSED_LOOP_CONTROL          8
-#define AXIS_STATE_LOCKIN_SPIN                  9
-#define AXIS_STATE_ENCODER_DIR_FIND             10
-
-// Simple example of irq safe queue (this is not multi-core safe)
-#define QUEUE_SIZE 128 // Must be power of 2
-static struct {
-    uint32_t pull_pos;
-    volatile uint32_t push_pos;
-    struct can2040_msg queue[QUEUE_SIZE];
-} MessageQueue;
-
-// Internal storage for can2040 module
-static struct can2040 cbus;
-
-// Main canbus callback (called from irq handler)
-static void
-can2040_cb(struct can2040 *cd, uint32_t notify, struct can2040_msg *msg)
-{
-    if (notify == CAN2040_NOTIFY_RX) {
-        // Add to queue
-        uint32_t push_pos = MessageQueue.push_pos;
-        uint32_t pull_pos = MessageQueue.pull_pos;
-        if ((push_pos + 1) % QUEUE_SIZE == pull_pos)
-            // No space in queue
-            return;
-        MessageQueue.queue[push_pos % QUEUE_SIZE] = *msg;
-        MessageQueue.push_pos = (push_pos + 1) % QUEUE_SIZE;
-    }
-}
-
-// PIO interrupt handler
-static void
-PIOx_IRQHandler(void)
-{
-    can2040_pio_irq_handler(&cbus);
-}
-
-// Initialize the can2040 module
-void
-canbus_setup(void)
-{
-    uint32_t pio_num = 0;
-    uint32_t sys_clock = SYS_CLK_HZ, bitrate = 250000;
-    uint32_t gpio_rx = 20, gpio_tx = 21;
-    //uint32_t gpio_rx = 26, gpio_tx = 27;
-
-    // Setup canbus
-    can2040_setup(&cbus, pio_num);
-    can2040_callback_config(&cbus, can2040_cb);
-
-    // Enable irqs
-    irq_set_exclusive_handler(PIO0_IRQ_0, PIOx_IRQHandler);
-    irq_set_priority(PIO0_IRQ_0, 1);
-    irq_set_enabled(PIO0_IRQ_0, 1);
-
-    // Start canbus
-    can2040_start(&cbus, sys_clock, bitrate, gpio_rx, gpio_tx);
-    
-    printf("CAN bus initialized: RX=GPIO%d, TX=GPIO%d, Baudrate=%d\n", 
-           gpio_rx, gpio_tx, bitrate);
-}
-
-// Send a CAN message with error checking
-bool send_can_message(uint32_t can_id, uint8_t dlc, uint8_t* data, const char* description, bool is_rtr)
-{
-    struct can2040_msg msg;
-    msg.id = can_id;
-
-    if (is_rtr){
-        msg.id |= CAN2040_ID_RTR;
-    }
-
-    memset(msg.data, 0, 8);
-    msg.dlc = dlc;
-    if (data && dlc > 0) {
-        memcpy(msg.data, data, dlc);
-    }
-    
-    int status = can2040_transmit(&cbus, &msg);
-    
-    printf("%s: CAN_ID=0x%03X, DLC=%d, Status=%d\n", 
-           description, can_id, dlc, status);
-    
-    if (dlc > 0 && data) {
-        printf("  Data: ");
-        for (int i = 0; i < dlc; i++) {
-            printf("%02X ", data[i]);
-        }
-        printf("\n");
-    }
-    
-    if (status != 0) {
-        printf("  ERROR: Transmission failed!\n");
-        return false;
-    }
-    
-    return true;
-}
 
 // Set ODrive axis state
 bool set_axis_state(uint8_t node_id, uint32_t axis_state)
@@ -261,8 +127,8 @@ main(void)
     set_axis_state(node_id, AXIS_STATE_IDLE);
     sleep_ms(1000);
     
-    printf("\nSetting axis to FULL_CALIBRATION_SEQUENCE state...\n");
-    set_axis_state(node_id, AXIS_STATE_FULL_CALIBRATION_SEQUENCE);
+    //printf("\nSetting axis to FULL_CALIBRATION_SEQUENCE state...\n");
+    //set_axis_state(node_id, AXIS_STATE_FULL_CALIBRATION_SEQUENCE);
     sleep_ms(10000);
     // Now try to set CLOSED_LOOP_CONTROL
     printf("\nAttempting to set CLOSED_LOOP_CONTROL...\n");
