+++ src/xdrive_canbus.h
@@ -1,0 +1,41 @@
+// ODrive CAN Commands
+#define ODR_CMD_HEARTBEAT           0x001
+#define ODR_CMD_ESTOP               0x002
+#define ODR_CMD_GET_MOTOR_ERROR     0x003
+#define ODR_CMD_GET_ENCODER_ERROR   0x004
+#define ODR_CMD_GET_SENSORLESS_ERROR 0x005
+#define ODR_CMD_SET_AXIS_NODE_ID    0x006
+#define ODR_CMD_SET_AXIS_STATE      0x007
+#define ODR_CMD_GET_ENCODER_ESTIMATES 0x009
+#define ODR_CMD_GET_ENCODER_COUNT   0x00A
+#define ODR_CMD_SET_CONTROLLER_MODES 0x00B
+#define ODR_CMD_SET_INPUT_POS       0x00C
+#define ODR_CMD_SET_INPUT_VEL       0x00D
+#define ODR_CMD_SET_INPUT_TORQUE    0x00E
+#define ODR_CMD_SET_VEL_LIMIT       0x00F
+#define ODR_CMD_START_ANTICOGGING   0x010
+#define ODR_CMD_SET_TRAJ_VEL_LIMIT  0x011
+#define ODR_CMD_SET_TRAJ_ACCEL_LIMITS 0x012
+#define ODR_CMD_SET_TRAJ_INERTIA    0x013
+#define ODR_CMD_GET_IQ              0x014
+#define ODR_CMD_GET_SENSORLESS_ESTIMATES 0x015
+#define ODR_CMD_REBOOT              0x016
+#define ODR_CMD_GET_VBUS_VOLTAGE    0x017
+#define ODR_CMD_CLEAR_ERRORS        0x018
+#define ODR_CMD_SET_LINEAR_COUNT    0x019
+#define ODR_CMD_SET_POS_GAIN        0x01A
+#define ODR_CMD_SET_VEL_GAINS       0x01B
+
+// ODrive Axis States
+#define AXIS_STATE_UNDEFINED                    0
+#define AXIS_STATE_IDLE                         1
+#define AXIS_STATE_STARTUP_SEQUENCE             2
+#define AXIS_STATE_FULL_CALIBRATION_SEQUENCE    3
+#define AXIS_STATE_MOTOR_CALIBRATION            4
+#define AXIS_STATE_SENSORLESS_CONTROL           5
+#define AXIS_STATE_ENCODER_INDEX_SEARCH         6
+#define AXIS_STATE_ENCODER_OFFSET_CALIBRATION   7
+#define AXIS_STATE_CLOSED_LOOP_CONTROL          8
+#define AXIS_STATE_LOCKIN_SPIN                  9
+#define AXIS_STATE_ENCODER_DIR_FIND             10
+
